Time: 296 ms (81.32%) | Memory: 62.7 MB (81.32%) - LeetSync

+++ 1557-check-if-a-string-contains-all-binary-codes-of-size-k/check-if-a-string-contains-all-binary-codes-of-size-k.cpp
@@ -21,6 +21,6 @@ public:
             hash = (hash + codes[s[i+k]]) % MOD;
             hashs.insert(hash);
         }
-        return hashs.size() == (long long)pow(2,k);
+        return hashs.size() == pow(2,k);
     }
 };
